+++ library/src/main/java/com/evernote/android/job/JobManager.java
@@ -406,9 +406,8 @@ public final class JobManager {
     }
 
     private boolean cancelInner(@Nullable Job job) {
-        if (job != null && !job.isFinished() && !job.isCanceled()) {
+        if (job != null && job.cancel(true)) {
             CAT.i("Cancel running %s", job);
-            job.cancel(true);
             return true;
         } else {
             return false;