+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix misleading log message for some internal improvements, see #391
 * Fix race condition when scheduling a job with `setUpdateCurrent(true)` where multiple jobs could have been scheduled, see #396
 * Fix bug where a daily job runs twice a day, see #406
+* Fix a bug where periodic jobs in the flex support mode weren't properly canceled while the job was running, see #407
 
 ## 1.2.4 (2018-02-08)
 * Add `scheduleAsync()` to the `DailyJob` class for scheduling daily jobs asynchronously to avoid IO operations on the main thread, see #371